+++ worker/bigraph/steps.py
@@ -23,8 +23,8 @@ except:
         'on installing Smoldyn.'
     )
 
-from simularium_utils import calculate_agent_radius, translate_data_object, write_simularium_file
-from shared import PROCESS_TYPES
+from worker.simularium_utils import calculate_agent_radius, translate_data_object, write_simularium_file
+from worker import APP_PROCESS_REGISTRY as PROCESS_TYPES
 
 CORE = PROCESS_TYPES
 process_registry = ProcessTypes().process_registry
@@ -569,3 +569,91 @@ def make_fallback_serializer_function() -> Callable:
         return serializer.serialize(obj)
     return default
 
+
+# -- Output data generators: -- #
+
+class OutputGenerator(Step):
+    config_schema = {
+        'input_file': 'string',
+        'context': 'string',
+    }
+
+    def __init__(self, config=None, core=APP_PROCESS_REGISTRY):
+        super().__init__(config, core)
+        self.input_file = self.config['input_file']
+        self.context = self.config.get('context')
+        if self.context is None:
+            raise ValueError("context (i.e., simulator name) must be specified in this processes' config.")
+
+    @abstractmethod
+    def generate(self, parameters: Optional[Dict[str, Any]] = None):
+        """Abstract method for generating output data upon which to base analysis from based on its origin.
+
+        This can be used for logic of any scope.
+        NOTE: args and kwargs are not defined in this function, but rather should be defined by the
+        inheriting class' constructor: i,e; start_time, etc.
+
+        Kwargs relate only to the given simulator api you are working with.
+        """
+        pass
+
+    def initial_state(self):
+        # base class method
+        return {
+            'output_data': {}
+        }
+
+    def inputs(self):
+        return {
+            'parameters': 'tree[any]'
+        }
+
+    def outputs(self):
+        return {
+            'output_data': 'tree[any]'
+        }
+
+    def update(self, state):
+        parameters = state.get('parameters') if isinstance(state, dict) else {}
+        data = self.generate(parameters)
+        return {'output_data': data}
+
+
+class TimeCourseOutputGenerator(OutputGenerator):
+    # NOTE: we include defaults here as opposed to constructor for the purpose of deliberate declaration within .json state representation.
+    config_schema = {
+        # 'input_file': 'string',
+        # 'context': 'string',
+        'start_time': {
+            '_type': 'integer',
+            '_default': 0
+        },
+        'end_time': {
+            '_type': 'integer',
+            '_default': 10
+        },
+        'num_steps': {
+            '_type': 'integer',
+            '_default': 100
+        },
+    }
+
+    def __init__(self, config=None, core=APP_PROCESS_REGISTRY):
+        super().__init__(config, core)
+        if not self.input_file.endswith('.xml'):
+            raise ValueError('Input file must be a valid SBML (XML) file')
+
+        self.start_time = self.config.get('start_time')
+        self.end_time = self.config.get('end_time')
+        self.num_steps = self.config.get('num_steps')
+        self.species_mapping = get_sbml_species_mapping(self.input_file)
+
+    def initial_state(self):
+        # TODO: implement this
+        pass
+
+    def generate(self, parameters: Optional[Dict[str, Any]] = None):
+        # TODO: add kwargs (initial state specs) here
+        executor = SBML_EXECUTORS[self.context]
+        data = executor(self.input_file, self.start_time, self.end_time, self.num_steps)
+        return data
