+++ worker/data_generator.py
@@ -1,17 +1,16 @@
-import os
+from abc import ABC, abstractmethod
 from tempfile import mkdtemp
+import traceback
+from pprint import pformat
 from typing import *
-from importlib import import_module
 
 import libsbml
-import numpy as np
-from kisao import AlgorithmSubstitutionPolicy
-from biosimulators_utils.config import Config
+from process_bigraph import Step, Composite
 
+from worker import APP_PROCESS_REGISTRY
 from worker.shared_worker import handle_exception
-from worker.data_model import BiosimulationsRunOutputData
 from worker.compatible import COMPATIBLE_UTC_SIMULATORS
-from worker.io_worker import read_report_outputs, normalize_smoldyn_output_path_in_root, make_dir, read_h5_reports
+from worker.io_worker import read_report_outputs, normalize_smoldyn_output_path_in_root, make_dir, read_h5_reports, get_sbml_species_mapping
 
 AMICI_ENABLED = True
 COPASI_ENABLED = True
@@ -46,17 +45,194 @@ except ImportError as e:
     PYSCES_ENABLED = False
 
 
-# def generate_smoldyn_simularium(smoldyn_configuration_file: str, output_dest_dir: str, use_json: bool = True, agent_params=None, box_size=None):
-#     # 1. make temp dir with config file written to it and set that as archive root
-#     temp_archive_root = mkdtemp()
-#     with open(smoldyn_configuration_file, 'r') as fh:
-#         smoldyn_config = fh.read()
+# class OutputGenerator(Step):
+#     config_schema = {
+#         'input_file': 'string',
+#         'context': 'string',
+#     }
 #
-#     with open(os.path.join(temp_archive_root, smoldyn_configuration_file.split('/')[-1]), 'w') as f:
-#         f.write(smoldyn_config)
+#     def __init__(self, config=None, core=APP_PROCESS_REGISTRY):
+#         super().__init__(config, core)
+#         self.input_file = self.config['input_file']
+#         self.context = self.config.get('context')
+#         if self.context is None:
+#             raise ValueError("context (i.e., simulator name) must be specified in this processes' config.")
 #
-#     return execute_simularium(working_dir=temp_archive_root, use_json=use_json, output_dir=output_dest_dir, agent_params=agent_params, box_size=box_size)
+#     @abstractmethod
+#     def generate(self, parameters: Optional[Dict[str, Any]] = None):
+#         """Abstract method for generating output data upon which to base analysis from based on its origin.
+#
+#         This can be used for logic of any scope.
+#         NOTE: args and kwargs are not defined in this function, but rather should be defined by the
+#         inheriting class' constructor: i,e; start_time, etc.
+#
+#         Kwargs relate only to the given simulator api you are working with.
+#         """
+#         pass
+#
+#     def initial_state(self):
+#         # base class method
+#         return {
+#             'output_data': {}
+#         }
+#
+#     def inputs(self):
+#         return {
+#             'parameters': 'tree[any]'
+#         }
+#
+#     def outputs(self):
+#         return {
+#             'output_data': 'tree[any]'
+#         }
+#
+#     def update(self, state):
+#         parameters = state.get('parameters') if isinstance(state, dict) else {}
+#         data = self.generate(parameters)
+#         return {'output_data': data}
+#
+#
+# class TimeCourseOutputGenerator(OutputGenerator):
+#     # NOTE: we include defaults here as opposed to constructor for the purpose of deliberate declaration within .json state representation.
+#     config_schema = {
+#         # 'input_file': 'string',
+#         # 'context': 'string',
+#         'start_time': {
+#             '_type': 'integer',
+#             '_default': 0
+#         },
+#         'end_time': {
+#             '_type': 'integer',
+#             '_default': 10
+#         },
+#         'num_steps': {
+#             '_type': 'integer',
+#             '_default': 100
+#         },
+#     }
+#
+#     def __init__(self, config=None, core=APP_PROCESS_REGISTRY):
+#         super().__init__(config, core)
+#         if not self.input_file.endswith('.xml'):
+#             raise ValueError('Input file must be a valid SBML (XML) file')
+#
+#         self.start_time = self.config.get('start_time')
+#         self.end_time = self.config.get('end_time')
+#         self.num_steps = self.config.get('num_steps')
+#         self.species_mapping = get_sbml_species_mapping(self.input_file)
+#
+#     def initial_state(self):
+#         # TODO: implement this
+#         pass
+#
+#     def generate(self, parameters: Optional[Dict[str, Any]] = None):
+#         # TODO: add kwargs (initial state specs) here
+#         executor = SBML_EXECUTORS[self.context]
+#         data = executor(self.input_file, self.start_time, self.end_time, self.num_steps)
+#         return data
 
+
+# register utc instance:
+# CORE.process_registry.register('time-course-output-generator', TimeCourseOutputGenerator)
+# register generic instance:
+# CORE.process_registry.register('output-generator', OutputGenerator)
+
+
+# -- functions related to generating time course output data (for verification and more) using the process-bigraph engine -- #
+
+def node_spec(_type: str, address: str, config: Dict[str, Any], inputs: Dict[str, Any], outputs: Dict[str, Any], name: str = None):
+    spec = {
+        '_type': _type,
+        'address': address,
+        'config': config,
+        'inputs': inputs,
+        'outputs': outputs
+    }
+
+    return {name: spec} if name else spec
+
+
+def step_node_spec(address: str, config: Dict[str, Any], inputs: Dict[str, Any], outputs: Dict[str, Any], name: str = None):
+    return node_spec(name=name, _type="step", address=address, config=config, inputs=inputs, outputs=outputs)
+
+
+def process_node_spec(address: str, config: Dict[str, Any], inputs: Dict[str, Any], outputs: Dict[str, Any], name: str = None):
+    return node_spec(name=name, _type="process", address=address, config=config, inputs=inputs, outputs=outputs)
+
+
+def time_course_node_spec(input_file: str, context: str, start_time: int, end_time: int, num_steps: int):
+    config = {
+        'input_file': input_file,
+        'start_time': start_time,
+        'end_time': end_time,
+        'num_steps': num_steps,
+        'context': context
+    }
+    return step_node_spec(
+        address='local:time-course-output-generator',
+        config=config,
+        inputs={
+            'parameters': [f'parameters_store_{context}']
+        },
+        outputs={
+            'output_data': [f'output_data_store_{context}']
+        }
+    )
+
+
+def generate_time_course_data(
+        input_fp: str,
+        start: int,
+        end: int,
+        steps: int,
+        simulators: List[str] = None,
+        parameters: Dict[str, Any] = None,
+        expected_results_fp: str = None,
+        out_dir: str = None
+) -> Dict[str, Dict[str, List[float]]]:
+    requested_sims = simulators or ["amici", "copasi", "pysces", "tellurium"]
+    simulation_spec = {
+        simulator: time_course_node_spec(
+            input_file=input_fp,
+            context=simulator,
+            start_time=start,
+            end_time=end,
+            num_steps=steps
+        ) for simulator in requested_sims
+    }
+    simulation = Composite({'state': simulation_spec, 'emitter': {'mode': 'all'}}, core=APP_PROCESS_REGISTRY)
+
+    input_filename = input_fp.split("/")[-1].split(".")[0]
+    if out_dir:
+        simulation.save(
+            filename=f'{input_filename}-initialization.json',
+            outdir=out_dir
+        )
+
+    # TODO: is there a better way to do this? (interval of one? Is that symbolic more than anything?)
+    if parameters:
+        simulation.update(parameters, 1)
+    else:
+        simulation.run(1)
+
+    if out_dir:
+        simulation.save(
+            filename=f'{input_filename}-update.json',
+            outdir=out_dir
+        )
+
+    output_data = {}
+    raw_data = simulation.gather_results()[('emitter',)]
+    for data in raw_data:
+        for data_key, data_value in data.items():
+            if data_key.startswith('output_data_store_'):
+                simulator = data_key.split('_')[-1]
+                output_data[simulator] = data_value
+
+    return output_data
+
+
+# -- direct simulator API wrappers --
 
 # TODO: should we return the actual data from memory, or that reflected in a smoldyn output txt file or both?
 def run_smoldyn(model_fp: str, duration: int, dt: float = None) -> Dict[str, Union[str, Dict[str, Union[float, List[float]]]]]:
@@ -132,43 +308,15 @@ def run_smoldyn(model_fp: str, duration: int, dt: float = None) -> Dict[str, Uni
             output_data = {'results_file': results_fp}
     except:
         error = handle_exception("Run Smoldyn")
+        logger.error(error)
         output_data = {'error': error}
 
     return output_data
 
 
-def get_sbml_species_mapping(sbml_fp: str):
-    """
-
-    Args:
-        - sbml_fp: `str`: path to the SBML model file.
-
-    Returns:
-        Dictionary mapping of {sbml_species_names(usually the actual observable name): sbml_species_ids(ids used in the solver)}
-    """
-    # read file
-    sbml_reader = libsbml.SBMLReader()
-    sbml_doc = sbml_reader.readSBML(sbml_fp)
-    sbml_model_object = sbml_doc.getModel()
-    # parse and handle names/ids
-    sbml_species_ids = []
-    for spec in sbml_model_object.getListOfSpecies():
-        spec_name = spec.name
-        if not spec_name:
-            spec.name = spec.getId()
-        if not spec.name == "":
-            sbml_species_ids.append(spec)
-    names = list(map(lambda s: s.name, sbml_species_ids))
-    species_ids = [spec.getId() for spec in sbml_species_ids]
-    return dict(zip(names, species_ids))
-
-
 def handle_sbml_exception() -> str:
-    import traceback
-    from pprint import pformat
     tb_str = traceback.format_exc()
-    error_message = pformat(f"SBML Simulation Error:\n{tb_str}")
-    
+    error_message = pformat(f"time-course-simulation-error:\n{tb_str}")
     return error_message
 
 
@@ -195,13 +343,13 @@ def run_sbml_pysces(sbml_fp: str, start, dur, steps):
             }
         except:
             error_message = handle_sbml_exception()
+            logger.error(error_message)
             return {"error": error_message}
     else:
         return OSError('Pysces is not properly installed in your environment.')
 
 
 def run_sbml_tellurium(sbml_fp: str, start, dur, steps):
-    # in the case that the start time is set to a point after the simulation begins
     result = None
     try:
         simulator = te.loadSBMLModel(sbml_fp)
@@ -222,6 +370,7 @@ def run_sbml_tellurium(sbml_fp: str, start, dur, steps):
             raise Exception('Tellurium: Could not generate results.')
     except:
         error_message = handle_sbml_exception()
+        logger.error(error_message)
         return {"error": error_message}
 
 
@@ -238,6 +387,7 @@ def run_sbml_copasi(sbml_fp, start, dur, steps):
         return data
     except:
         error_message = handle_sbml_exception()
+        logger.error(error_message)
         return {"error": error_message}
 
 
@@ -283,199 +433,12 @@ def run_sbml_amici(sbml_fp: str, start, dur, steps):
         return results
     except:
         error_message = handle_sbml_exception()
+        logger.error(error_message)
         return {"error": error_message}
 
 
-def _generate_biosimulator_utc_outputs(omex_fp: str, output_root_dir: str, simulators: List[str] = None, alg_policy="same_framework") -> Dict:
-    """Generate the outputs of the standard UTC simulators Copasi, Tellurium, and Amici from the
-        biosimulators interface (exec_sedml_docs_in_combine_archive).
-    """
-    make_dir(output_root_dir)
-
-    output_data = {}
-    sims = simulators or ['amici', 'copasi', 'tellurium']
-    sim_config = Config(
-        LOG=False,
-        ALGORITHM_SUBSTITUTION_POLICY=AlgorithmSubstitutionPolicy[alg_policy.upper()],
-        VERBOSE=False)
-    for sim in sims:
-        sim_output_dir = os.path.join(output_root_dir, f'{sim}_outputs')
-        make_dir(sim_output_dir)
-        try:
-            module = import_module(name=f'biosimulators_{sim}.core')
-            exec_func = getattr(module, 'exec_sedml_docs_in_combine_archive')
-            sim_output_dir = os.path.join(output_root_dir, f'{sim}_outputs')
-            if not os.path.exists(sim_output_dir):
-                os.mkdir(sim_output_dir)
-            # execute simulator-specific simulation
-            exec_func(archive_filename=omex_fp, out_dir=sim_output_dir, config=sim_config)
-            report_path = os.path.join(sim_output_dir, 'reports.h5')
-
-            sim_data = read_report_outputs(report_path)
-            data = sim_data.to_dict() if isinstance(sim_data, BiosimulationsRunOutputData) else sim_data
-            output_data[sim] = data
-        except Exception as e:
-            import traceback
-            tb_str = traceback.format_exc()
-            error_message = (
-                f"An unexpected error occurred while processing your request:\n"
-                f"Error Type: {type(e).__name__}\n"
-                f"Error Details: {str(e)}\n"
-                f"Traceback:\n{tb_str}"
-            )
-            output_data[sim] = error_message
-
-    return output_data
-
-
-def generate_biosimulator_utc_outputs(omex_fp: str, output_root_dir: str, simulators: list[str] = None, alg_policy="same_framework") -> dict:
-    """Generate the outputs of the standard UTC simulators Copasi, Tellurium, and Amici from the
-        biosimulators interface (exec_sedml_docs_in_combine_archive).
-    """
-    make_dir(output_root_dir)
-
-    output_data = {}
-    sims = simulators or ['amici', 'copasi', 'tellurium']  # , 'pysces']
-    sim_config = Config(
-        LOG=False,
-        ALGORITHM_SUBSTITUTION_POLICY=AlgorithmSubstitutionPolicy[alg_policy.upper()],
-        VERBOSE=False)
-    for sim in sims:
-        sim_output_dir = os.path.join(output_root_dir, f'{sim}_outputs')
-        make_dir(sim_output_dir)
-        try:
-            module = import_module(name=f'biosimulators_{sim}.core')
-            exec_func = getattr(module, 'exec_sedml_docs_in_combine_archive')
-            sim_output_dir = os.path.join(output_root_dir, f'{sim}_outputs')
-            if not os.path.exists(sim_output_dir):
-                os.mkdir(sim_output_dir)
-
-            # execute simulator-specific simulation
-            exec_func(
-                archive_filename=omex_fp,
-                out_dir=sim_output_dir,
-                config=sim_config if not sim == "pysces" else None
-            )
-            report_path = os.path.join(sim_output_dir, 'reports.h5')
-
-            sim_data = read_h5_reports(report_path)
-            data = sim_data.to_dict() if isinstance(sim_data, BiosimulationsRunOutputData) else sim_data
-            output_data[sim] = data
-        except:
-            import traceback
-            tb_str = traceback.format_exc()
-            error_message = (
-                f"Traceback:\n{tb_str}"
-            )
-            output_data[sim] = {'error': error_message}
-
-    return output_data
-
-
-# TODO: add Vcell and pysces here
+# TODO: add vcell and masspy here
 SBML_EXECUTORS = dict(zip(
     [data[0] for data in COMPATIBLE_UTC_SIMULATORS],
     [run_sbml_amici, run_sbml_copasi, run_sbml_pysces, run_sbml_tellurium]
 ))
-
-
-def generate_sbml_utc_outputs(sbml_fp: str, start: int, dur: int, steps: int, simulators: list[str] = None, truth: str = None) -> dict:
-    # TODO: add VCELL and pysces here
-    output = {}
-    sbml_species_ids = list(get_sbml_species_mapping(sbml_fp).keys())
-    simulators = simulators or ['amici', 'copasi', 'tellurium', 'pysces']
-    all_output_ids = []
-    for simulator in simulators:
-        results = {}
-        simulator = simulator.lower()
-        simulation_executor = SBML_EXECUTORS[simulator]
-        sim_result = simulation_executor(sbml_fp=sbml_fp, start=start, dur=dur, steps=steps)
-
-        # case: simulation execution was successful
-        if "error" not in sim_result.keys():
-            # add to all shared names
-            all_output_ids.append(list(sim_result.keys()))
-
-            # iterate over sbml_species_ids to index output data
-            for species_id in sbml_species_ids:
-                if species_id in sim_result.keys():
-                    output_vals = sim_result[species_id]
-                    if isinstance(output_vals, np.ndarray):
-                        output_vals = output_vals.tolist()
-                    results[species_id] = output_vals
-        else:
-            # case: simulation had an error
-            results = sim_result
-
-        # set the simulator output
-        output[simulator] = results
-
-    # get the commonly shared output ids
-    final_output = {}
-    shared_output_ids = min(all_output_ids)
-    for simulator_name in output.keys():
-        sim_data = {}
-        for spec_id in output[simulator_name].keys():
-            if spec_id in shared_output_ids:
-                sim_data[spec_id] = output[simulator_name][spec_id]
-            elif spec_id == "error":
-                sim_data["error"] = output[simulator_name][spec_id]
-                
-        final_output[simulator_name] = sim_data
-
-    # handle expected outputs
-    if truth is not None:
-        final_output['ground_truth'] = {}
-        report_results = read_report_outputs(truth)
-        report_data = report_results.to_dict()['data'] if isinstance(report_results, BiosimulationsRunOutputData) else {}
-        for datum in report_data:
-            spec_name = datum['dataset_label']
-            if not spec_name.lower() == 'time':
-                spec_data = datum['data']
-                final_output['ground_truth'][spec_name] = spec_data
-
-    return final_output
-
-
-# def sbml_output_stack(spec_name: str, output):
-#     stack = []
-#     for simulator_name in output.keys():
-#         spec_data = output[simulator_name].get(spec_name)
-#         if spec_data is not None:
-#             stack.append(spec_data)
-#     return stack
-
-
-def sbml_output_stack(spec_name: str, output):
-    # 2. in output_stack: return {simname: output}
-    stack = {}
-    for simulator_name in output.keys():
-        spec_data = output[simulator_name].get(spec_name)
-        if isinstance(spec_data, str):
-            data = None
-        else:
-            data = spec_data
-
-        stack[simulator_name] = data
-
-    return stack
-
-
-def get_output_stack(spec_name: str, outputs):
-    return sbml_output_stack(spec_name=spec_name, output=outputs)
-
-
-def _get_report_output_stack(outputs: dict, spec_id: str):
-    output_stack = []
-    for sim_name in outputs.keys():
-        sim_data = outputs[sim_name]['data']
-        for data_index, data in enumerate(sim_data):
-            data_id = data['dataset_label']
-            if data_id == spec_id:
-                # print(spec_id, data_id)
-                output_stack.append(sim_data[data_index]['data'])
-            else:
-                pass
-    return np.stack(output_stack)
-
-
